refactor: 이름 변경에 따른 수정
이름 변경과 책임 분리:
RequestParser : 파싱 역할
RequestProcessor: 요청에 대해 httpRequest, httpResponse 를 만들고 HttpRequestDispatcher를 호출하는 역할
