+++ src/main/java/webserver/RequestProcessor.java
@@ -6,25 +6,20 @@ import http.constant.HttpStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import util.MimeType;
+import util.RequestParser;
 import util.exception.InvalidRequestLineSyntaxException;
 
 import java.io.*;
-import java.util.ArrayList;
-import java.util.List;
 
-public class RequestParser {
-    private static final Logger logger = LoggerFactory.getLogger(RequestParser.class);
-
-    public RequestParser() {
+public class RequestProcessor {
+    private static final Logger logger = LoggerFactory.getLogger(RequestProcessor.class);
+    public RequestProcessor() {
     }
 
-    public void parse(InputStream in, DataOutputStream dos) throws IOException {
+    public void process(InputStream in, DataOutputStream dos) throws IOException {
         try{
-            List<String> request = logAndReturnRequest(in);
+            HttpRequest httpRequest = RequestParser.parse(in);
 
-            String[] requestLine = resolveRequestLine(request.get(0));
-
-            HttpRequest httpRequest =  new HttpRequest(requestLine[0], requestLine[1], requestLine[2], request.subList(1, request.size()));
             HttpResponse httpResponse = new HttpResponse(dos);
 
             HttpRequestDispatcher httpRequestDispatcher = new HttpRequestDispatcher(httpRequest, httpResponse);
@@ -36,36 +31,7 @@ public class RequestParser {
         }
     }
 
-    private List<String> logAndReturnRequest(InputStream in) throws IOException {
-        List<String> request = new ArrayList<>();
 
-        DataInputStream dis = new DataInputStream(in);
-        StringBuilder sb = new StringBuilder();
-        int i;
-        while((i = dis.read()) != -1) {
-            char c = (char) i;
-            sb.append(c);
-            if (dis.available() == 0) break;
-        }
-
-        sb.trimToSize();
-        String[] tokens = sb.toString().split("\r\n");
-        for (String token : tokens) {
-            logger.debug("request: {}", token);
-            request.add(token);
-        }
-        return request;
-    }
-
-    private String[] resolveRequestLine(String requestLine) {
-        String[] tokens =  requestLine.split(" ");
-
-        if (tokens.length != 3) {
-            throw new InvalidRequestLineSyntaxException("");
-        }
-
-        return tokens;
-    }
 
     private void errorResponse(DataOutputStream dos, HttpStatus httpStatus, Exception e) throws IOException {
         logger.error(e.getMessage());
